Small fixes in API and organize project structure
