+++ backend/go/api/api_helm_applications.go
@@ -6,58 +6,48 @@
  * API version: 0.0.1
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
-package swagger
+package api
 
 import (
 	"net/http"
 )
 
-func CreateClusterResource(w http.ResponseWriter, r *http.Request) {
+func GetHelmRelease(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
 
-func CreateNamespacedResource(w http.ResponseWriter, r *http.Request) {
+func GetHelmReleaseHistory(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
 
-func DeleteClusterResource(w http.ResponseWriter, r *http.Request) {
+func InstallHelmRelease(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
 
-func DeleteNamespacedResource(w http.ResponseWriter, r *http.Request) {
+func ListHelmReleases(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
 
-func GetClusterResource(w http.ResponseWriter, r *http.Request) {
+func ListHelmReleasesInNamespace(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
 
-func GetNamespacedResource(w http.ResponseWriter, r *http.Request) {
+func RollbackHelmRelease(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
 
-func ListClusterResources(w http.ResponseWriter, r *http.Request) {
+func UninstallHelmRelease(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
 
-func ListNamespacedResources(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-	w.WriteHeader(http.StatusOK)
-}
-
-func UpdateClusterResource(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-	w.WriteHeader(http.StatusOK)
-}
-
-func UpdateNamespacedResource(w http.ResponseWriter, r *http.Request) {
+func UpdateHelmRelease(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
