+++ backend/go/api/api_kubernetes_resources.go
@@ -6,48 +6,58 @@
  * API version: 0.0.1
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
-package swagger
+package api
 
 import (
 	"net/http"
 )
 
-func GetHelmRelease(w http.ResponseWriter, r *http.Request) {
+func CreateClusterResource(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
 
-func GetHelmReleaseHistory(w http.ResponseWriter, r *http.Request) {
+func CreateNamespacedResource(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
 
-func InstallHelmRelease(w http.ResponseWriter, r *http.Request) {
+func DeleteClusterResource(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
 
-func ListHelmReleases(w http.ResponseWriter, r *http.Request) {
+func DeleteNamespacedResource(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
 
-func ListHelmReleasesInNamespace(w http.ResponseWriter, r *http.Request) {
+func GetClusterResource(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
 
-func RollbackHelmRelease(w http.ResponseWriter, r *http.Request) {
+func GetNamespacedResource(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
 
-func UninstallHelmRelease(w http.ResponseWriter, r *http.Request) {
+func ListClusterResources(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
 
-func UpdateHelmRelease(w http.ResponseWriter, r *http.Request) {
+func ListNamespacedResources(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+	w.WriteHeader(http.StatusOK)
+}
+
+func UpdateClusterResource(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+	w.WriteHeader(http.StatusOK)
+}
+
+func UpdateNamespacedResource(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
