+++ backend/go/api/routers.go
@@ -6,10 +6,11 @@
  * API version: 0.0.1
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
-package swagger
+package api
 
 import (
 	"fmt"
+	"github.com/ZPI-2024-25/KubernetesUserManager/go/common"
 	"net/http"
 	"strings"
 
@@ -30,7 +31,7 @@ func NewRouter() *mux.Router {
 	for _, route := range routes {
 		var handler http.Handler
 		handler = route.HandlerFunc
-		handler = Logger(handler, route.Name)
+		handler = common.Logger(handler, route.Name)
 
 		router.
 			Methods(route.Method).
@@ -85,126 +86,126 @@ var routes = Routes{
 	Route{
 		"GetHelmRelease",
 		strings.ToUpper("Get"),
-		"/api/v1/api/helm/releases/{namespace}/{releaseName}",
+		"/api/v1/helm/releases/{namespace}/{releaseName}",
 		GetHelmRelease,
 	},
 
 	Route{
 		"GetHelmReleaseHistory",
 		strings.ToUpper("Get"),
-		"/api/v1/api/helm/releases/{namespace}/{releaseName}/history",
+		"/api/v1/helm/releases/{namespace}/{releaseName}/history",
 		GetHelmReleaseHistory,
 	},
 
 	Route{
 		"InstallHelmRelease",
 		strings.ToUpper("Post"),
-		"/api/v1/api/helm/releases/{namespace}",
+		"/api/v1/helm/releases/{namespace}",
 		InstallHelmRelease,
 	},
 
 	Route{
 		"ListHelmReleases",
 		strings.ToUpper("Get"),
-		"/api/v1/api/helm/releases",
+		"/api/v1/helm/releases",
 		ListHelmReleases,
 	},
 
 	Route{
 		"ListHelmReleasesInNamespace",
 		strings.ToUpper("Get"),
-		"/api/v1/api/helm/releases/{namespace}",
+		"/api/v1/helm/releases/{namespace}",
 		ListHelmReleasesInNamespace,
 	},
 
 	Route{
 		"RollbackHelmRelease",
 		strings.ToUpper("Post"),
-		"/api/v1/api/helm/releases/{namespace}/{releaseName}/rollback",
+		"/api/v1/helm/releases/{namespace}/{releaseName}/rollback",
 		RollbackHelmRelease,
 	},
 
 	Route{
 		"UninstallHelmRelease",
 		strings.ToUpper("Delete"),
-		"/api/v1/api/helm/releases/{namespace}/{releaseName}",
+		"/api/v1/helm/releases/{namespace}/{releaseName}",
 		UninstallHelmRelease,
 	},
 
 	Route{
 		"UpdateHelmRelease",
 		strings.ToUpper("Put"),
-		"/api/v1/api/helm/releases/{namespace}/{releaseName}",
+		"/api/v1/helm/releases/{namespace}/{releaseName}",
 		UpdateHelmRelease,
 	},
 
 	Route{
 		"CreateClusterResource",
 		strings.ToUpper("Post"),
-		"/api/v1/api/{resourceType}",
+		"/api/v1/k8s/non-namespaced/{resourceType}",
 		CreateClusterResource,
 	},
 
 	Route{
 		"CreateNamespacedResource",
 		strings.ToUpper("Post"),
-		"/api/v1/api/{resourceType}/{namespace}",
+		"/api/v1/k8s/namespaced/{resourceType}/{namespace}",
 		CreateNamespacedResource,
 	},
 
 	Route{
 		"DeleteClusterResource",
 		strings.ToUpper("Delete"),
-		"/api/v1/api/{resourceType}/{resourceName}",
+		"/api/v1/k8s/non-namespaced/{resourceType}/{resourceName}",
 		DeleteClusterResource,
 	},
 
 	Route{
 		"DeleteNamespacedResource",
 		strings.ToUpper("Delete"),
-		"/api/v1/api/{resourceType}/{namespace}/{resourceName}",
+		"/api/v1/k8s/namespaced/{resourceType}/{namespace}/{resourceName}",
 		DeleteNamespacedResource,
 	},
 
 	Route{
 		"GetClusterResource",
 		strings.ToUpper("Get"),
-		"/api/v1/api/{resourceType}/{resourceName}",
+		"/api/v1/k8s/non-namespaced/{resourceType}/{resourceName}",
 		GetClusterResource,
 	},
 
 	Route{
 		"GetNamespacedResource",
 		strings.ToUpper("Get"),
-		"/api/v1/api/{resourceType}/{namespace}/{resourceName}",
+		"/api/v1/k8s/namespaced/{resourceType}/{namespace}/{resourceName}",
 		GetNamespacedResource,
 	},
 
 	Route{
 		"ListClusterResources",
 		strings.ToUpper("Get"),
-		"/api/v1/api/{resourceType}",
+		"/api/v1/k8s/non-namespaced/{resourceType}",
 		ListClusterResources,
 	},
 
 	Route{
 		"ListNamespacedResources",
 		strings.ToUpper("Get"),
-		"/api/v1/api/{resourceType}/{namespace}",
+		"/api/v1/k8s/namespaced/{resourceType}/{namespace}",
 		ListNamespacedResources,
 	},
 
 	Route{
 		"UpdateClusterResource",
 		strings.ToUpper("Put"),
-		"/api/v1/api/{resourceType}/{resourceName}",
+		"/api/v1/k8s/non-namespaced/{resourceType}/{resourceName}",
 		UpdateClusterResource,
 	},
 
 	Route{
 		"UpdateNamespacedResource",
 		strings.ToUpper("Put"),
-		"/api/v1/api/{resourceType}/{namespace}/{resourceName}",
+		"/api/v1/k8s/namespaced/{resourceType}/{namespace}/{resourceName}",
 		UpdateNamespacedResource,
 	},
