+++ backend/go/common/logger.go
@@ -6,14 +6,26 @@
  * API version: 0.0.1
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
-package swagger
+package common
 
-// Authentication tokens
-type AuthTokens struct {
-	// JWT access token
-	AccessToken string `json:"accessToken,omitempty"`
-	// Refresh token
-	RefreshToken string `json:"refreshToken,omitempty"`
-	// Access token expiration time in seconds
-	ExpiresIn int32 `json:"expiresIn,omitempty"`
+import (
+	"log"
+	"net/http"
+	"time"
+)
+
+func Logger(inner http.Handler, name string) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		start := time.Now()
+
+		inner.ServeHTTP(w, r)
+
+		log.Printf(
+			"%s %s %s %s",
+			r.Method,
+			r.RequestURI,
+			name,
+			time.Since(start),
+		)
+	})
 }
