--- backend/go/model_metadata.go
@@ -1,27 +1,0 @@
-/*
- * KubernetesUserManager - API
- *
- * This is a backend API server documentation for KubernetesUserManager  Some useful links: - [Jira](https://samuelus.atlassian.net/jira/software/projects/ZPI/boards/4) - [Confluence](https://samuelus.atlassian.net/wiki/spaces/ZPI/overview)
- *
- * API version: 0.0.1
- * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
- */
-package swagger
-
-import (
-	"time"
-)
-
-// Standard Kubernetes metadata
-type Metadata struct {
-	// Name of the resource
-	Name string `json:"name,omitempty"`
-	// Namespace of the resource
-	Namespace string `json:"namespace,omitempty"`
-	// Key-value pairs for labels
-	Labels map[string]string `json:"labels,omitempty"`
-	// Key-value pairs for annotations
-	Annotations map[string]string `json:"annotations,omitempty"`
-	// Timestamp when the resource was created
-	CreationTimestamp time.Time `json:"creationTimestamp,omitempty"`
-}
