+++ backend/go/models/model_auth_callback_body.go
@@ -6,7 +6,7 @@
  * API version: 0.0.1
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
-package swagger
+package models
 
 type AuthCallbackBody struct {
 	// Authorization code.
