+++ backend/go/models/model_auth_tokens.go
@@ -6,11 +6,14 @@
  * API version: 0.0.1
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
-package swagger
+package models
 
-// User session details and privileges
-type UserStatus struct {
-	User *User `json:"user,omitempty"`
-	// List of user permissions
-	Permissions []string `json:"permissions,omitempty"`
+// Authentication tokens
+type AuthTokens struct {
+	// JWT access token
+	AccessToken string `json:"accessToken,omitempty"`
+	// Refresh token
+	RefreshToken string `json:"refreshToken,omitempty"`
+	// Access token expiration time in seconds
+	ExpiresIn int32 `json:"expiresIn,omitempty"`
 }
