+++ backend/go/models/model_error.go
@@ -6,9 +6,14 @@
  * API version: 0.0.1
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
-package swagger
+package models
 
-type IdPermissionsBody struct {
-	// List of permissions.
-	Permissions []string `json:"permissions,omitempty"`
+// Error response
+type ModelError struct {
+	// Error code
+	Code int32 `json:"code,omitempty"`
+	// Error message
+	Message string `json:"message,omitempty"`
+	// Additional error details
+	Details *interface{} `json:"details,omitempty"`
 }
