+++ backend/go/models/model_helm_release.go
@@ -6,7 +6,7 @@
  * API version: 0.0.1
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
-package swagger
+package models
 
 import (
 	"time"
@@ -25,5 +25,7 @@ type HelmRelease struct {
 	// Last update timestamp
 	Updated time.Time `json:"updated,omitempty"`
 	// Revision number of the release
-	Revision int32 `json:"revision,omitempty"`
+	Revision string `json:"revision,omitempty"`
+	// App version of the release
+	AppVersion string `json:"app_version,omitempty"`
 }
