+++ backend/go/models/model_helm_release_history.go
@@ -6,7 +6,7 @@
  * API version: 0.0.1
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
-package swagger
+package models
 
 import (
 	"time"
@@ -22,6 +22,8 @@ type HelmReleaseHistory struct {
 	Status string `json:"status,omitempty"`
 	// Chart version used
 	Chart string `json:"chart,omitempty"`
+	// App version of the release
+	AppVersion string `json:"app_version,omitempty"`
 	// Description of the revision
 	Description string `json:"description,omitempty"`
 }
