+++ backend/go/models/model_id_permissions_body.go
@@ -6,26 +6,9 @@
  * API version: 0.0.1
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
-package swagger
+package models
 
-import (
-	"log"
-	"net/http"
-	"time"
-)
-
-func Logger(inner http.Handler, name string) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		start := time.Now()
-
-		inner.ServeHTTP(w, r)
-
-		log.Printf(
-			"%s %s %s %s",
-			r.Method,
-			r.RequestURI,
-			name,
-			time.Since(start),
-		)
-	})
+type IdPermissionsBody struct {
+	// List of permissions.
+	Permissions []string `json:"permissions,omitempty"`
 }
