+++ backend/go/models/model_release_name_rollback_body.go
@@ -6,9 +6,9 @@
  * API version: 0.0.1
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
-package swagger
+package models
 
-type AuthRefreshBody struct {
-	// Refresh token.
-	RefreshToken string `json:"refreshToken,omitempty"`
+type ReleaseNameRollbackBody struct {
+	// Version number to rollback to.
+	Version int32 `json:"version,omitempty"`
 }
