+++ backend/go/models/model_resource.go
@@ -6,7 +6,7 @@
  * API version: 0.0.1
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
-package swagger
+package models
 
 // Generic Kubernetes resource
 type Resource struct {
@@ -14,8 +14,8 @@ type Resource struct {
 	ApiVersion string `json:"apiVersion,omitempty"`
 	// Kind of the resource
 	Kind string `json:"kind,omitempty"`
-
-	Metadata *Metadata `json:"metadata,omitempty"`
+	// Specification of the metadata
+	Metadata *interface{} `json:"metadata,omitempty"`
 	// Specification of the resource
 	Spec *interface{} `json:"spec,omitempty"`
 	// Current status of the resource
