+++ backend/go/models/model_status.go
@@ -6,14 +6,14 @@
  * API version: 0.0.1
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
-package swagger
+package models
 
-// Error response
-type ModelError struct {
-	// Error code
-	Code int32 `json:"code,omitempty"`
-	// Error message
+// Status response after an operation
+type Status struct {
+	// Status of the operation (e.g., Success, Failure)
+	Status string `json:"status,omitempty"`
+	// Message providing additional information
 	Message string `json:"message,omitempty"`
-	// Additional error details
-	Details *interface{} `json:"details,omitempty"`
+	// HTTP status code
+	Code int32 `json:"code,omitempty"`
 }
