+++ backend/go/models/model_user_status.go
@@ -6,14 +6,11 @@
  * API version: 0.0.1
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
-package swagger
+package models
 
-// Status response after an operation
-type Status struct {
-	// Status of the operation (e.g., Success, Failure)
-	Status string `json:"status,omitempty"`
-	// Message providing additional information
-	Message string `json:"message,omitempty"`
-	// HTTP status code
-	Code int32 `json:"code,omitempty"`
+// User session details and privileges
+type UserStatus struct {
+	User *User `json:"user,omitempty"`
+	// List of user permissions
+	Permissions []string `json:"permissions,omitempty"`
 }
